Extract composable functions out of the fragment
